OTT-548: Removed non-selected model groups from req.ext.prebid.floors
