+++ floors/rule_test.go
@@ -217,3 +217,86 @@ func TestShouldSkipFloors(t *testing.T) {
 	}
 
 }
+
+func TestSelectFloorModelGroup(t *testing.T) {
+	floorExt := &openrtb_ext.PriceFloorRules{Data: &openrtb_ext.PriceFloorData{
+		SkipRate: 30,
+		ModelGroups: []openrtb_ext.PriceFloorModelGroup{{
+			ModelWeight:  50,
+			SkipRate:     10,
+			ModelVersion: "Version 1",
+			Schema:       openrtb_ext.PriceFloorSchema{Fields: []string{"mediaType", "size", "domain"}},
+			Values: map[string]float64{
+				"banner|300x250|www.website.com": 1.01,
+				"banner|300x250|*":               2.01,
+				"banner|300x600|www.website.com": 3.01,
+				"banner|300x600|*":               4.01,
+				"banner|728x90|www.website.com":  5.01,
+				"banner|728x90|*":                6.01,
+				"banner|*|www.website.com":       7.01,
+				"banner|*|*":                     8.01,
+				"*|300x250|www.website.com":      9.01,
+				"*|300x250|*":                    10.01,
+				"*|300x600|www.website.com":      11.01,
+				"*|300x600|*":                    12.01,
+				"*|728x90|www.website.com":       13.01,
+				"*|728x90|*":                     14.01,
+				"*|*|www.website.com":            15.01,
+				"*|*|*":                          16.01,
+			}, Default: 0.01},
+			{
+				ModelWeight:  25,
+				SkipRate:     20,
+				ModelVersion: "Version 2",
+				Schema:       openrtb_ext.PriceFloorSchema{Fields: []string{"mediaType", "size", "domain"}},
+				Values: map[string]float64{
+					"banner|300x250|www.website.com": 1.01,
+					"banner|300x250|*":               2.01,
+					"banner|300x600|www.website.com": 3.01,
+					"banner|300x600|*":               4.01,
+					"banner|728x90|www.website.com":  5.01,
+					"banner|728x90|*":                6.01,
+					"banner|*|www.website.com":       7.01,
+					"banner|*|*":                     8.01,
+					"*|300x250|www.website.com":      9.01,
+					"*|300x250|*":                    10.01,
+					"*|300x600|www.website.com":      11.01,
+					"*|300x600|*":                    12.01,
+					"*|728x90|www.website.com":       13.01,
+					"*|728x90|*":                     14.01,
+					"*|*|www.website.com":            15.01,
+					"*|*|*":                          16.01,
+				}, Default: 0.01},
+		}}}
+
+	tt := []struct {
+		name         string
+		floorExt     *openrtb_ext.PriceFloorRules
+		ModelVersion string
+		fn           func(int) int
+	}{
+		{
+			name:         "banner|300x250|www.website.com",
+			floorExt:     floorExt,
+			ModelVersion: "Version 2",
+			fn:           func(i int) int { return 5 },
+		},
+		{
+			name:         "banner|300x600|www.website.com",
+			floorExt:     floorExt,
+			ModelVersion: "Version 1",
+			fn:           func(i int) int { return 55 },
+		},
+	}
+
+	for _, tc := range tt {
+		t.Run(tc.name, func(t *testing.T) {
+			selectFloorModelGroup(tc.floorExt.Data.ModelGroups, tc.fn)
+
+			if !reflect.DeepEqual(tc.floorExt.Data.ModelGroups[0].ModelVersion, tc.ModelVersion) {
+				t.Errorf("Floor Model Version mismatch error: \nreturn:\t%v\nwant:\t%v", tc.floorExt.Data.ModelGroups[0].ModelVersion, tc.ModelVersion)
+			}
+
+		})
+	}
+}
