+++ openrtb_ext/floors.go
@@ -7,6 +7,7 @@ type PriceFloorSchema struct {
 type PriceFloorModelGroup struct {
 	Currency     string             `json:"currency,omitempty"`
 	ModelWeight  int                `json:"modelweight,omitempty"`
+	DebugWeight  int                `json:"debugweight,omitempty"` // Added for Debug purpose, shall be removed
 	ModelVersion string             `json:"modelversion,omitempty"`
 	SkipRate     int                `json:"skiprate,omitempty"`
 	Schema       PriceFloorSchema   `json:"schema,omitempty"`
@@ -40,5 +41,5 @@ type PriceFloorRules struct {
 	Data        *PriceFloorData        `json:"data,omitempty"`
 	Enforcement *PriceFloorEnforcement `json:"enforcement,omitempty"`
 	Enabled     *bool                  `json:"enabled,omitempty"`
-	Skipped     *bool
+	Skipped     *bool                  `json:"skipped,omitempty"`
 }
